% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionOccurrenceTests.R
\name{createConditionOccurrenceTests}
\alias{createConditionOccurrenceTests}
\title{Run the condition occurrence tests}
\usage{
createConditionOccurrenceTests()
}
\description{
Run the condition occurrence tests
}
