% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationPeriodTests.R
\name{createObservationPeriodTests}
\alias{createObservationPeriodTests}
\title{Run the observation period tests}
\usage{
createObservationPeriodTests()
}
\description{
Run the observation period tests
}
