% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcedureOccurrenceTests.R
\name{createProcedureOccurrenceTests}
\alias{createProcedureOccurrenceTests}
\title{Run the procedure occurrence tests}
\usage{
createProcedureOccurrenceTests()
}
\description{
Run the procedure occurrence tests
}
