% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitTests.R
\name{createTests}
\alias{createTests}
\title{Function that groups the test cases in a logical way to write the sql file}
\usage{
createTests()
}
\description{
Function that groups the test cases in a logical way to write the sql file
}
